Move testing fakes to testing.go
