+++ notify/grafana_alertmanager_test.go
@@ -313,16 +313,6 @@ func TestPutAlert(t *testing.T) {
 	}
 }
 
-type fakeNotifier struct{}
-
-func (f *fakeNotifier) Notify(_ context.Context, _ ...*types.Alert) (bool, error) {
-	return true, nil
-}
-
-func (f *fakeNotifier) SendResolved() bool {
-	return true
-}
-
 func TestGrafanaAlertmanager_setReceiverMetrics(t *testing.T) {
 	fn := &fakeNotifier{}
 	integrations := []*notify.Integration{
@@ -411,46 +401,4 @@ func TestSilenceCleanup(t *testing.T) {
 		require.NoError(t, err)
 		return len(found) == 2
 	}, 6*time.Second, 150*time.Millisecond)
-}
-
-type FakeConfig struct {
-}
-
-func (f *FakeConfig) DispatcherLimits() DispatcherLimits {
-	panic("implement me")
-}
-
-func (f *FakeConfig) InhibitRules() []*InhibitRule {
-	// TODO implement me
-	panic("implement me")
-}
-
-func (f *FakeConfig) MuteTimeIntervals() []MuteTimeInterval {
-	// TODO implement me
-	panic("implement me")
-}
-
-func (f *FakeConfig) ReceiverIntegrations() (map[string][]Integration, error) {
-	// TODO implement me
-	panic("implement me")
-}
-
-func (f *FakeConfig) RoutingTree() *Route {
-	// TODO implement me
-	panic("implement me")
-}
-
-func (f *FakeConfig) Templates() *Template {
-	// TODO implement me
-	panic("implement me")
-}
-
-func (f *FakeConfig) Hash() [16]byte {
-	// TODO implement me
-	panic("implement me")
-}
-
-func (f *FakeConfig) Raw() []byte {
-	// TODO implement me
-	panic("implement me")
 }
